Explicitly untie and close the output file handle because ParseXS was holding the file handle open, preventing the compiler from opening it. [Randy W. Sims]
Added an '--output FILENAME' flag to xsubpp and changed ParseXS to use
the named file in the #line directives when the output file has an
extension other than '.c' (i.e. '.cpp'). [Randy W. Sims]

Added conditional definition of the PERL_UNUSED_VAR macro to the
output file in case it's not already defined for backwards
compatibility with pre-5.8 versions of perl. (Not sure if this is the
best solution.) [Randy W. Sims]