Make INCLUDE_COMMAND semi-portable.
At the very least, arguments to the command must be quoted on VMS
to preserve case.  The double quotes then have to be escaped to
be parseable as a string literal by the C preprocessor for the #line
directive.  The result is still dubious, as it is in no way "the
presumed name of the source file" as the C standard says it must
be, but this does get it passing tests.

This change also stops pasting the directory onto the command
since it's a command, not a path, and for the INCLUDE handler,
it properly assembles the path using File::Spec->catfile().